chore(gcal): relocate self-contained Google utilities out of packages/backend/src/sync/
gcal.context.ts (a general common/ file, not part of the legacy engine)
imported getGcalClient and isMissingGoogleRefreshToken from inside
packages/backend/src/sync/, which is the final wholesale-deletion target
for this whole cleanup effort. Both functions turned out to have zero
imports from anywhere else in sync/ — they were only physically located
there, not architecturally tied to it. Same for pruneGoogleDataAndNotifyRevoked,
called unconditionally from error.express.handler.ts's generic Express
error handler (not delegation-gated, reachable regardless of which engine
is active).

Relocated all three into packages/backend/src/common/services/gcal/,
their natural home alongside gcal.service.ts/gcal.context.ts/gcal.utils.ts:
- gcal.client.ts (getGcalClient, getGAuthClientForUser) moved as-is, plus
  its test file.
- isMissingGoogleRefreshToken (was its own tiny google-sync.errors.ts)
  folded directly into gcal.utils.ts, next to the sibling Google-error
  predicates (isInvalidGoogleToken, isGoogleError, isFullSyncRequired)
  it belongs with.
- pruneGoogleDataAndNotifyRevoked (was google-sync.revoked.ts) moved to
  a new google-revoked.util.ts.

Updated every import path across the 8 files that referenced the old
locations (error.express.handler.ts, gcal.context.ts, sync.controller.ts,
compass-to-google.event-propagation.ts, google-sync.service.ts,
google-watch.service.ts, google-watch-repair.service.ts,
google-watch-maintenance.planner.ts). No behavior change — this is a
pure relocation, nothing was deleted or rewritten.

Investigated but deliberately left alone: gcal.service.ts's
encodeChannelToken/watchCalendars/watchEvents have no caller outside
sync/'s own google-watch.service.ts, so they die alongside sync/ in the
final deletion pass with nothing left dangling — no relocation pressure.
error.express.handler.ts's repairGoogleCalendarSync and getSyncByToken
calls are genuinely legacy-specific for the same reason. user.service.ts's
syncRecords/googleWatchService account-deletion sweep is NOT dead code
(real data-privacy stakes, not delegation-gated) and needs its own
careful investigation — tracked separately, not rushed into this PR.
