+++ packages/backend/src/common/services/gcal/gcal.client.test.ts
@@ -3,7 +3,7 @@ import { GaxiosError } from "gaxios";
 import { ObjectId } from "mongodb";
 import { type Schema_User } from "@core/types/user.types";
 import { UserError } from "@backend/common/errors/user/user.errors";
-import { getGcalClient } from "@backend/sync/services/google-sync/gcal.client";
+import { getGcalClient } from "@backend/common/services/gcal/gcal.client";
 import * as userQueries from "@backend/user/queries/user.queries";
 import { describe, expect, it, spyOn } from "bun:test";
