--- packages/backend/src/sync/services/google-sync/google-sync.errors.ts
@@ -1,11 +1,0 @@
-import { BaseError } from "@core/errors/errors.base";
-import { UserError } from "@backend/common/errors/user/user.errors";
-
-export const isMissingGoogleRefreshToken = (
-  error: unknown,
-): error is BaseError => {
-  return (
-    error instanceof BaseError &&
-    error.description === UserError.MissingGoogleRefreshToken.description
-  );
-};
