+++ packages/backend/src/sync/services/google-sync/google-sync.service.ts
@@ -12,11 +12,11 @@ import {
   type GoogleRequestContext,
 } from "@backend/common/services/gcal/gcal.context";
 import { isInvalidGoogleToken } from "@backend/common/services/gcal/gcal.utils";
+import { pruneGoogleDataAndNotifyRevoked } from "@backend/common/services/gcal/google-revoked.util";
 import { createConcurrencyLimiter } from "@backend/common/util/concurrency-limiter.util";
 import { DEFAULT_GCAL_EVENTS_PER_PAGE } from "@backend/event/google-event-sync.service";
 import { sseServer } from "@backend/servers/sse/sse.server";
 import compassToGoogleBackfill from "@backend/sync/services/event-propagation/compass-to-google/compass-to-google-backfill";
-import { pruneGoogleDataAndNotifyRevoked } from "@backend/sync/services/google-sync/google-sync.revoked";
 import {
   createSyncImport,
   type SyncImport,
